Moved /poem and /add_samples to /analyze/poem etc...
