+++ run/analyze/__init__.py
@@ -1,0 +1,7 @@
+"""Initialize the analyze blueprint."""
+
+from flask import Blueprint
+
+analyze = Blueprint('analyze', __name__)
+
+from . import views
