+++ run/main/views.py
@@ -1,10 +1,10 @@
 """Route definitions for the main blueprint."""
 
 import os
-from flask import render_template, redirect, url_for
+from flask import render_template
 from . import main
 from .. import db
-from ..models import Meter, Poet, Poem
+from ..models import Poem
 
 import markdown
 
@@ -31,32 +31,3 @@ def about():
     return render_template("about.html",
                            contents=ABOUT_CONTENTS,
                            )
-
-
-@main.route("/poem/<keyword>")
-def poem(keyword):
-    """Define the poem route."""
-    # if the poems table does not exist, 404 the route
-    if "poems" not in db.engine.table_names():
-        return render_template('404.html'), 404
-
-    # retrieve the requested poem if it exists
-    poem = Poem.query.filter_by(keyword=keyword).first_or_404()
-
-    return render_template("poem.html",
-                           title=poem.title,
-                           poet=poem.author.name,
-                           meter=poem.meter.name,
-                           poem=poem.HTML)
-
-
-@main.route("/add_samples")
-def add_samples():
-    """Define the add_samples route."""
-    try:
-        Meter.insert_samples()
-        Poet.insert_samples()
-        Poem.insert_samples()
-    except Exception:
-        return render_template('404.html'), 404
-    return redirect(url_for('.home'))
