+++ tests/test_flask/test_analyze_blueprint/test_add_samples_view.py
@@ -6,9 +6,9 @@ def test_dropped_db(client):
     """Check behavior when the db has no tables."""
     db.drop_all()
     assert b'Sorry' in client.get('/').data
-    assert "404" in client.get('/poem/Flea').status
+    assert "404" in client.get('/analyze/poem/Flea').status
     # /add_samples route should not cause server error when db has no tables
-    assert "404" in client.get('/add_samples').status
+    assert "404" in client.get('/analyze/add_samples').status
 
 
 def test_add_samples_route(client):
@@ -17,20 +17,20 @@ def test_add_samples_route(client):
     assert "poems" in db.engine.table_names()
     # The database has no poems in it
     assert b'Sorry' in client.get('/').data
-    assert "404" in client.get('/poem/Flea').status
+    assert "404" in client.get('/analyze/poem/Flea').status
 
     # /add_samples route injects sample poems into database
-    client.get('/add_samples')
+    client.get('/analyze/add_samples')
     assert b'Flea' in client.get('/').data
-    response = client.get('/poem/Flea')
+    response = client.get('/analyze/poem/Flea')
     assert b'Flea' in response.data
     assert "200" in response.status
-    assert "404" in client.get('/poem/foo').status
+    assert "404" in client.get('/analyze/poem/foo').status
 
     # Makes sure doing it twice doesn't break anything.
-    client.get('/add_samples')
+    client.get('/analyze/add_samples')
     assert b'Flea' in client.get('/').data
-    response = client.get('/poem/Flea')
+    response = client.get('/analyze/poem/Flea')
     assert b'Flea' in response.data
     assert "200" in response.status
-    assert "404" in client.get('/poem/foo').status
+    assert "404" in client.get('/analyze/poem/foo').status
